Initial commit.
Please find my extension of the "external_feedback"-block in this
commit. After all the time I finally found the time to open source it,
but it definately needs to be extended to appeal for a larger audience.

For now, please be careful and take a look at the README.md and pay
attention at its limitations. In any case, please contact me by e-mail
by mailing to any adress at jo2k.de.

Of course, no warranty or liability for any damage can be given, as this
piece of software is only provided by best-will and best-effort.

Have fun.

-- Jonathan Liebers